Fix Elephants' factor implementation
This commit changes the implementation part of the metric to use
a percentage calculation instead of a percentile calculation.

The 50th percentile divides the observations in two groups of the same
size, sorted by value, and then provides the value that corresponds
to the observation(s) placed in the position(s) used to split them.
This not represents the 50% of contributions but the value around
what we have equal number of observations on each side. With this,
the Elephant Factor will be always half of the total number of
observations.

Signed-off-by: Alberto Pérez García-Plaza <alpgarcia@bitergia.com>