Update references to the Elephant Factor metric.
Two unrelated references were removed. The paper "The Blind Men and
the Elephant" does not use the term Elephant Factor to measure the
dependency on a set of companies. Same with the second removed reference
in the blog philslade.blogspot.com. I've replaced those two with
the slide deck in OSCON 2015 by Jesus G. Barahona, when he presented
the factor for the first time. The last reference added includes a
more complete explanation of the factor by Bitergia.

Signed-off-by: Luis Cañas-Díaz <lcanas@bitergia.com>